Moved logic of building runnable to runnable-builder.
commit-id:87f37143
