+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
     "crates/cairo-lang-plugins",
     "crates/cairo-lang-proc-macros",
     "crates/cairo-lang-project",
+    "crates/cairo-lang-runnable-utils",
     "crates/cairo-lang-runner",
     "crates/cairo-lang-semantic",
     "crates/cairo-lang-sierra",
