+++ crates/cairo-lang-runnable-utils/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "cairo-lang-runnable-utils"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license-file.workspace = true
+description = "Helpers for creating cairo runnable artifact."
+
+[dependencies]
+cairo-lang-casm = { path = "../cairo-lang-casm", version = "~2.8.4" }
+cairo-lang-sierra = { path = "../cairo-lang-sierra", version = "~2.8.4" }
+cairo-lang-sierra-ap-change = { path = "../cairo-lang-sierra-ap-change", version = "~2.8.4" }
+cairo-lang-sierra-gas = { path = "../cairo-lang-sierra-gas", version = "~2.8.4" }
+cairo-lang-sierra-to-casm = { path = "../cairo-lang-sierra-to-casm", version = "~2.8.4" }
+cairo-lang-sierra-type-size = { path = "../cairo-lang-sierra-type-size", version = "~2.8.4" }
+cairo-lang-utils = { path = "../cairo-lang-utils", version = "~2.8.4" }
+cairo-vm.workspace = true
+itertools = { workspace = true, default-features = true }
+num-bigint = { workspace = true }
+thiserror.workspace = true
+
+[dev-dependencies]
