+++ crates/cairo-lang-runner/Cargo.toml
@@ -13,12 +13,10 @@ ark-secp256r1.workspace = true
 
 cairo-lang-casm = { path = "../cairo-lang-casm", version = "~2.8.4" }
 cairo-lang-lowering = { path = "../cairo-lang-lowering", version = "~2.8.4" }
+cairo-lang-runnable-utils = { path = "../cairo-lang-runnable-utils", version = "~2.8.4" }
 cairo-lang-sierra = { path = "../cairo-lang-sierra", version = "~2.8.4" }
-cairo-lang-sierra-ap-change = { path = "../cairo-lang-sierra-ap-change", version = "~2.8.4" }
-cairo-lang-sierra-gas = { path = "../cairo-lang-sierra-gas", version = "~2.8.4" }
 cairo-lang-sierra-generator = { path = "../cairo-lang-sierra-generator", version = "~2.8.4" }
 cairo-lang-sierra-to-casm = { path = "../cairo-lang-sierra-to-casm", version = "~2.8.4" }
-cairo-lang-sierra-type-size = { path = "../cairo-lang-sierra-type-size", version = "~2.8.4" }
 cairo-lang-starknet = { path = "../cairo-lang-starknet", version = "~2.8.4" }
 cairo-lang-utils = { path = "../cairo-lang-utils", version = "~2.8.4" }
 cairo-vm.workspace = true
