+++ scripts/release_crates.sh
@@ -34,6 +34,7 @@ CRATES_TO_PUBLISH=(
     cairo-lang-compiler
     cairo-lang-starknet-classes
     cairo-lang-starknet
+    cairo-lang-runnable-utils
     cairo-lang-runner
     cairo-lang-test-plugin
     cairo-lang-test-runner
